fix(ci): update tooltip and rust test layout
